\name{zfGenes}
\alias{zfGenes}
\alias{zebrafishRNASeq}
\docType{data}
\title{
  Zebrafish RNA-Seq Experimental Data from Ferreira et al. (2014)
 }
\description{
Gene-level read counts from the RNA-Seq experiment of Ferreira et
al. (2014). The data correspond to RNA libraries for three pairs of
gallein-treated and control embryonic zebrafish cell pools. Please see
the vignette for details on pre-processing.
}
\usage{data(zfGenes)}
\source{
  A citation for the dataset can be generated using
  \code{citation("zebrafishRNASeq")}. 
}
\keyword{datasets}
